Struct glowygraph::render2::Renderer

pub struct Renderer<'a> {
    // some fields omitted
}

Methods

impl<'a> Renderer<'a>

A Renderer is tied to the lifetime of the glium Display and making one builds a GLSL program internally.

fn new(display: &'a Display) -> Self

Make a new Renderer from a glium::Display.

fn render_nodes<S>(&self, target: &mut S, transform: [[f32; 3]; 3], nodes: &[Node]) where S: Surface

Take a series of nodes and draw them in parallel on the GPU.

fn render_edges_round<S>(&self, target: &mut S, transform: [[f32; 3]; 3], edges: &[Node]) where S: Surface

Take a series of lines (edges) and draw them in parallel on the GPU.

These will have round ends.

fn render_edges_flat<S>(&self, target: &mut S, transform: [[f32; 3]; 3], edges: &[Node]) where S: Surface

Take a series of lines (edges) and draw them in parallel on the GPU.

These will have flat ends.

fn render_qbeziers_round<S>(&self, target: &mut S, transform: [[f32; 3]; 3], qbeziers: &[QBezier]) where S: Surface

Take a series of triangles (quadratic bezier curves) and draw them in parallel on the GPU.

These will have round ends.

fn render_qbeziers_flat<S>(&self, target: &mut S, transform: [[f32; 3]; 3], qbeziers: &[QBezier]) where S: Surface

Take a series of triangles (quadratic bezier curves) and draw them in parallel on the GPU.

These will have flat ends.