Struct glowygraph::Renderer

pub struct Renderer<'a> {
    // some fields omitted
}

Methods

impl<'a> Renderer<'a>

A Renderer is tied to the lifetime of the glium Display and making one builds a GLSL program internally.

fn new(display: &'a Display) -> Self

Make a new Renderer from a glium::Display.

fn render_nodes(&self, modelview: &[[f32; 4]; 4], projection: &[[f32; 4]; 4], nodes: &[Node])

Take a modelview matrix, projection matrix, and a series of nodes and draw them in parallel on the GPU.